servicedb/requirements.txt
# To ensure app dependencies are ported from your virtual environment/host machine into your container, run 'pip freeze > requirements.txt' in the terminal to overwrite this file
flask==2.0.2
gunicorn==20.1.0
uvicorn[standard]
fastapi
pika
pydantic
SQLAlchemy
psycopg2-binary
python-dotenv
